+++ src/collective/blicca/footerblocks/i18n.py
@@ -1,4 +1,5 @@
 """Message factory for the collective.blicca.footerblocks i18n domain."""
+
 from zope.i18nmessageid import MessageFactory