+++ src/collective/blicca/footerblocks/setuphandlers.py
@@ -1,4 +1,5 @@
 """Setup handlers for collective.blicca.footerblocks."""
+
 from Products.CMFPlone.interfaces import INonInstallable
 from zope.interface import implementer